requirements.txt
Flask==3.1.0
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.6
pdf2image==1.17.0
python-dotenv==1.0.1
langchain-ollama==0.2.3
boto3==1.36.17
flask-cors==5.0.0